CIV-4321 - refactor and add claimant package/folder
