+++ src/main/common/form/models/claim/claimant/claimantPartyTypeSelection.ts
@@ -1,5 +1,5 @@
 import {IsDefined} from 'class-validator';
-import {CounterpartyType} from '../../../../common/models/counterpartyType';
+import {CounterpartyType} from 'models/counterpartyType';
 
 export class ClaimantPartyTypeSelection {
   @IsDefined({message: 'ERRORS.VALID_CHOOSE'})
